+++ docs/source/package_reference/cli.mdx
@@ -205,7 +205,7 @@ Arguments for options ran inside the TPU.
 * `--command_file` (`str`) -- The path to the file containing the commands to run on the pod on startup.
 * `--command` (`str`) -- A command to run on the pod. Can be passed multiple times.
 * `--install_accelerate` (`bool`) -- Whether to install accelerate on the pod. Defaults to False.
-* `--accelerate_version` (`str`) -- The version of accelerate to install on the pod. If not specified, will use the                        latest pypi version. Specify 'dev' to install from GitHub.
+* `--accelerate_version` (`str`) -- The version of accelerate to install on the pod. If not specified, will use the latest pypi version. Specify 'dev' to install from GitHub.
 * `--debug` (`bool`) -- If set, will print the command that would be run instead of running it.
 
 ## accelerate test
